PPPPPPP

+++ SalesAndInentoryWeb_Application/SalesAndInentoryWeb_Application/ViewModel/PartyDetails.cs
@@ -22,7 +22,7 @@ namespace SalesAndInentoryWeb_Application.ViewModel
         public string Barcode { get; set; }
         public double Received  { get; set; }
         public double RemainingBal { get; set; }
-        public int TaxAmount1 { get; set; }
+        public float TaxAmount1 { get; set; }
         public string Feild3 { get; set; }
         public double CalTotal { get; set; }
         public string Status { get; set; }
